feat(solver): implementación de CalculadorEnvyFreeness (#28)
En este PR se agregó una implementación `CalculadorEnvyFreeness` de la
interfaz `ICalculadoraFitness`.
Se implementó el método `ExtraerAsignacion` en `Individuo`.

+++ src/Solver/Fitness/ICalculadoraFitness.cs
@@ -4,6 +4,6 @@ namespace Solver.Fitness
 {
     internal interface ICalculadoraFitness
     {
-        int CalcularFitness(Individuo individuo, InstanciaProblema instanciaProblema);
+        decimal CalcularFitness(Individuo individuo, InstanciaProblema instanciaProblema);
     }
 }
